Separate nginx and mariadb in different networks
I have replaced the deleted project example with a new one located in
the "web application" folder.

In docker-compose file one can specify a hostname, which is very
convenient.

I have set "restart: always" so that the containers now always restart.
This means that the container will always start on daemon startup,
regardless of the current state of the container.

The nginx and mariadb services now use different networks for better
isolation.